backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.3
sentence-transformers==2.2.2
scikit-learn==1.3.0